合并查询接口

+++ zcc-cust/src/main/java/com/wensheng/zcc/cust/module/vo/CustCountVo.java
@@ -5,12 +5,12 @@ import lombok.Data;
 
 @Data
 public class CustCountVo {
-  private Integer allPersonCount;
-  private Integer allCmpycount;
-  private Integer updatePersonCount;
-  private Integer updateCmpycount;
-  private Integer creatPersonCount;
-  private Integer creatCmpycount;
-  private Integer tradePersonCount;
-  private Integer tradeCmpycount;
+  private Long allPersonCount;
+  private Long allCmpycount;
+  private Long updatePersonCount;
+  private Long updateCmpycount;
+  private Long creatPersonCount;
+  private Long creatCmpycount;
+  private Long tradePersonCount;
+  private Long tradeCmpycount;
 }
